Compiling on OSF. A couple of yet unsolved problems here: ld outputs warning "weak symbol multiply defined" and there's an artifact after compilation -- file "so_locations".
git-svn-id: https://svn.code.sf.net/p/stemkit/code@114 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/util/Object.h
@@ -30,7 +30,7 @@ namespace sk {
 	virtual const Class getClass() const;
 
 	/// Returns a unique object id.
-	virtual unsigned int getId() const;
+	virtual long long getId() const;
 
 	/// Returns a string representation of the object. By default, it
 	/// returns the object's class and the objects's id.
